+++ tests/decorators/test_class_decorators.py
@@ -1,4 +1,4 @@
-from pedantic.decorators.class_decorators import trace_class, timer_class
+from pedantic.decorators.class_decorators import trace_class
 
 
 def test_trace_class():
@@ -17,19 +17,3 @@ def test_trace_class():
     m = MyClass.generator()
     m.double(b=42)
 
-
-def test_timer_class():
-    @timer_class
-    class MyClass:
-        def __init__(self, s: str) -> None:
-            self.s = s
-
-        def double(self, b: int) -> str:
-            return self.s + str(b)
-
-        @staticmethod
-        def generator() -> 'MyClass':
-            return MyClass(s='generated')
-
-    m = MyClass.generator()
-    m.double(b=42)